Move gen xml v040 tests
